<refactoring> Command separate
 - read, write, command, factory
